refactor(select): add util to get disabled keys from items
